+++ dlc-gradle-plugin/src/main/java/io/domainlifecycles/plugin/DlcGradlePlugin.java
@@ -146,6 +146,7 @@ public class DlcGradlePlugin implements Plugin<Project> {
                 task.getStreamUpload().set(domainModelUploadTaskConfigurationExtension.getStreamUpload().convention(false));
                 task.getStaticAnalysisCacheSize().set(domainModelUploadTaskConfigurationExtension.getStaticAnalysisCacheSize()
                     .convention(DomainCallsAnalyzerImpl.DEFAULT_CACHE_SIZE));
+                task.getStaticAnalysisPackages().set(domainModelUploadTaskConfigurationExtension.getStaticAnalysisPackages());
                 task.getClassesDirs().from(totalClassesDirs);
                 task.getClasspath().from(totalClasspath);
             });